requirements.txt
xgboost~=2.0.0
pandas~=2.1.1
jupyter
matplotlib~=3.8.0
scikit-learn~=1.3.1
tensorflow~=2.14.0
PyYAML~=6.0.1
wandb
joblib~=1.3.2
seaborn~=0.13.0
numpy~=1.26.0
astral
black~=24.1
isort~=5.11
pre-commit~=3.6
pytest~=8.0